공격 속도 추가

+++ Source/ALegendaryMine/DataTable/MonsterDataTable.h
@@ -50,7 +50,7 @@ public:
 		int32 Speed;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ItemDataTable")
-		int32 AttackSpeed;
+		float AttackSpeed;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ItemDataTable")
 		int32 AttackDistance;
